Remove patch version of Python
This was presumably added to placate Dependabot.
Specifying the patch version is no longer needed as a workaround for
this.

See opensafely-core/job-server@f2059f8d8d5e14526a1468ad736687ce2537e2dc.

Additionally, currently, Dependabot fails with:

> Dependabot does not support your Python version

presumably because the patch version is old.

So this change should fix this.