chore: Only tag versions on `main` branch
See opensafely-core/ehrql@3e55492e9c1b537fb5057f19f11f53a713fbae46 for
more context around this.

In this workflow, we currently only run on pushes to main, but this acts
as an extra guard in case that changes. The tagging can become very
broken, with tags to non-release branches occurring.