feat: Add `GithubRepo.get_topics` method
- `get_topics` calls the repo endpoint to get the repo's topics.
The repo endpoint might have been called previously to retrieve the
repo's name and description, so it is possible that by calling
`get_topics` we would be calling the endpoint twice. This can be
avoided by setting `use_cache=True` when instantiating the
`GithubClient` for use.

- An alternative implementation would be to always retrieve topics along
with the repo's name and description, but this would require modifying
the API of the `GithubRepo` class and its `get_repo_details` method.
Changing the API seems like a large drawback, especially since topics
are not always needed.
Calling the endpoint a second time only when topics are needed and the
GithubClient is not using the cache seems like a reasonable trade-off.

- Topics are required by opensafely-core/actions-registry#433.